Refactor CalendarLayout to reduce complexity
