+++ frontend/src/components/calendar/CalendarDetailView.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export function CalendarDetailView({ selectedDetailEvent }: { selectedDetailEvent: Record<string, any> | null }) {
+  return (
+    <section aria-label="일정 상세" className="rounded-2xl border border-border bg-card p-5 shadow-sm">
+      <h3 className="text-lg font-bold">{selectedDetailEvent ? `${selectedDetailEvent.title} 상세` : '일정 상세'}</h3>
+      <dl className="mt-4 grid gap-4 md:grid-cols-2">
+        <div className="rounded-xl border border-border bg-background p-4">
+          <dt className="text-xs font-black text-muted-foreground">원본 계정</dt>
+          <dd className="mt-2 text-sm font-bold">{selectedDetailEvent ? `${selectedDetailEvent.source} · 충돌 토큰 확인` : '표시 중인 원본 없음'}</dd>
+        </div>
+        <div className="rounded-xl border border-border bg-background p-4">
+          <dt className="text-xs font-black text-muted-foreground">충돌 제어</dt>
+          <dd className="mt-2 text-sm font-bold">ETag / If-Match 필요 시 server-authoritative 검증</dd>
+        </div>
+      </dl>
+    </section>
+  );
+}
