+++ frontend/src/components/calendar/types.ts
@@ -1,0 +1,29 @@
+export type CalendarWritebackIntentResponse = {
+  workspace_id: string;
+  target_source_id: string;
+  protocol: string;
+  writeback_mode: 'customer_owned';
+  requires_if_match: boolean;
+  if_match: string | null;
+  provenance: Record<string, string>;
+  audit_event: string;
+  provider_write_executed: boolean;
+  status: string;
+  runner_request_id: string | null;
+  provider_status: number | null;
+  error_code: string | null;
+  retry_item_uid?: string | null;
+};
+
+export type CalendarWritebackSource = {
+  source_id: string;
+  provider: string;
+  protocol: 'caldav' | 'carddav' | 'webdav' | 'local';
+  owner_id: string;
+  organization_id: string | null;
+  capabilities: string[];
+  writeback_enabled: boolean;
+  etag: string | null;
+};
+
+export type WritebackStatus = 'idle' | 'loading' | 'success' | 'no_source' | 'conflict' | 'auth' | 'error';
